Set dtype to float32 in reference metric

+++ tests/unittests/segmentation/test_hausdorff_distance.py
@@ -50,7 +50,7 @@ _inputs = _Input(preds=preds, target=target)
 # Wrapper that converts to numpy to avoid Torch-to-numpy functional issues
 def torch_skimage_hausdorff_distance(p: torch.Tensor, t: torch.Tensor) -> torch.Tensor:
     out = skimage_hausdorff_distance(p.numpy(), t.numpy())
-    return torch.tensor([out])
+    return torch.tensor([out], dtype=torch.float32)
 
 
 @pytest.mark.parametrize(
